Update and rename cr_rotating.py to cr_rotating_de.py
